+++ src/main/java/com/authlete/jaxrs/ClientRegistrationRequestHandler.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019 Authlete, Inc.
+ * Copyright (C) 2019-2021 Authlete, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -263,6 +263,10 @@ public class ClientRegistrationRequestHandler extends BaseHandler
             case BAD_REQUEST:
                 // 400 Bad Request
                 return ResponseUtil.badRequest(content);
+
+            case UNAUTHORIZED:
+                // 401 Unauthorized
+                return ResponseUtil.unauthorized(content, null);
 
             case CREATED:
                 // 201 Created